Поработал с замечаниями
