--- src/main/kotlin/Printable.kt
@@ -1,3 +1,0 @@
-interface Printable {
-    fun getItem(): String
-}
